ADNCD-1948: Add workaround for all PRs being rebuilt after GitHub migration
On initial load of a PR material (handleGetLatestRevision), store all
seen PRs, not just the one we're building. For subsequent loads
(handleLatestRevisionSince) behave the same as before, and only store
the PR we're building.

Also tried fancier variants where you can configure the behavior per PR
material, even allowing to clear the state and rebuild everything.
However, there seems to be a GoCD caching bug such that PR material
config changes in the UI are properly displayed, but not actually
propagated to the PR plugin. It keeps using the old config until
GoCD is restarted, or the PR material (and due to dependency the
referencing config repo) is removed and readded.