ADNCD-2084: Fix handleLatestRevisionSince not returning new revisions
The issue was introduced with the upgrade of git-cmd to 2.0, where the
getRevisionsSince method was changed to compare the old commit to the
remote branch, instead of comparing it to the HEAD of the checkout.
Therefore the plugin doesn't properly detect revisions from latest PR
commit (HEAD) to last seen commit.

To fix this, reintroduced the old getRevisionSince method as
getRevisionsUntilHeadSince to our ExtendedGitCmdHelper class. It
requires a lot of duplication from GitCmdHelper because most of its
helper methods are private.

The alternative is to downgrade git-cmd to 1.4 again, but it would be
nice to keep the other fixes and upgrades of git-cmd.